rm uncessary code
