--- src/dense/svds/SVD.jl
@@ -1,1 +1,0 @@
-include("gesvd/gesvd2.jl")
