--- src/dense/svds/gesvd/magmaDgesvd.jl
@@ -1,13 +1,0 @@
-# magma_dgesvd wrapper
-function magmaDgesvd(jobu, jobvt, m, n, A, lda, s, U,
-	 ldu, VT, ldvt, work, lwork, info)
-    ccall((:magma_dgesvd, libmagma),
-               Cint,
-               (Cint, Cint, Cint, Cint, PtrOrCuPtr{Cdouble},
-			    Cint, PtrOrCuPtr{Cdouble}, PtrOrCuPtr{Cdouble},
-				Cint, PtrOrCuPtr{Cdouble}, Cint,
-				PtrOrCuPtr{Cdouble}, Cint, PtrOrCuPtr{Cint}),
-
-                jobu, jobvt, m, n, A, lda, s, U, ldu, VT, ldvt,
-			    work, lwork, info)
-end
