--- src/utilities/Utilities.jl
@@ -1,1 +1,0 @@
-include("init.jl")
