--- src/utils/Utilities.jl
@@ -1,11 +1,0 @@
-module Utilities
-
-# import some common definitions from MAGMA
-import MAGMA.libmagma
-
-# Initialization and Finalization
-export magmaInit, magmaFinalize
-
-include("init.jl")
-
-end  # module Utilities
