--- src/utils/init.jl
@@ -1,11 +1,0 @@
-#export magmaInit, magmaFinalize
-
-# magma_init
-function magmaInit()
-	ccall((:magma_init, libmagma),Cint,())
-end
-
-# magma_finalize
-function magmaFinalize()
-	ccall((:magma_finalize, libmagma),Cint,())
-end
